cell-ai-app/requirements.txt
streamlit
torch
torchvision
opencv-python
pillow
numpy
pandas
matplotlib